GIAIC-Learn-Agentic-AI-/OpenRouter/pyproject.toml
[project]
name = "lit-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.72.4",
    "openai>=1.86.0",
    "python-dotenv>=1.1.0",
]
